+++ src/Services/DispatchWorkflowService.php
@@ -593,13 +593,8 @@ class DispatchWorkflowService
                                 $data[$index]['email'] = explode(',', $emailPlaceHolderValue);
                             }
 
-                            $data[$index]['email'] = $this->workflowService->removeUnsubscribedEmails(
-                                $data[$index]['email'],
-                                [
-                                    'workflow_id' => $this->workflowId,
-                                    'job_workflow_id' => $jobWorkflowId,
-                                    'campaign_type' => $action['payload']['campaignType'] ?? null,
-                                ]
+                            $data[$index]['email'] = $this->workflowService->getAllowedEmails(
+                                $data[$index]['email']
                             );
                         }
                     }
